+++ iib/iib-scripts/stop_node.sh
@@ -17,5 +17,7 @@ echo "----------------------------------------"
 echo "Stopping node $NODENAME..."
 endmqm $MQ_QMGR_NAME
 # Kill the rsyslogd process
-sudo kill $(cat /var/run/syslogd.pid | awk '{ print $1 }')
+if [ -e /var/run/syslogd.pid ]; then
+	sudo kill $(cat /var/run/syslogd.pid | awk '{ print $1 }')
+fi
 exit
